+++ kitty/boss.py
@@ -673,6 +673,15 @@ class Boss:
         if tab:
             self.confirm_tab_close(tab)
 
+    @ac('tab', 'Close all the tabs in the current OS window other than the currently active tab')
+    def close_other_tabs_in_os_window(self) -> None:
+        tm = self.active_tab_manager
+        if tm is not None and len(tm.tabs) > 1:
+            active_tab = self.active_tab
+            for tab in tm:
+                if tab is not active_tab:
+                    self.close_tab(tab)
+
     def confirm(
         self, msg: str,  # can contain newlines and ANSI formatting
         callback: Callable[..., None],  # called with True or False and *args